requirements.txt
PyPDF2==3.0.1
python-docx==1.1.2
streamlit==1.45.1
textract==1.6.3
langchain==0.3.25
langchain-core==0.3.60
langchain-mistralai==0.2.10